Well then...
Well that last update still gave me the same results but I tried running away and that reset them. So I removed that extra bit of code to test if that's still the case without it. If so then that'll have to do for now and might push it onto the realm to see how it is on there.